Added `ExpressionDataAccessCommandFactory`
